+++ threadmill-store-memory/src/test/java/com/hemju/threadmill/store/memory/FatalErrorBoundaryTest.java
@@ -41,7 +41,7 @@ import com.hemju.threadmill.core.schedule.RecurringMaterializer;
 import com.hemju.threadmill.core.serialization.JsonJobSerializer;
 import com.hemju.threadmill.core.spec.JobArgument;
 import com.hemju.threadmill.core.spec.JobSpec;
-import com.hemju.threadmill.test.ForwardingJobStore;
+import com.hemju.threadmill.core.store.ForwardingJobStore;
 
 class FatalErrorBoundaryTest {